The IO.for_fd use breaks fd lifecycle.
In ruby the runtime determines fd number and it is not possible to
request a higher fd number to avoid recently used FDs. When closing an
IO the fd is removed from the IO and it is forever closed. However, if
duplicate IO exists the fd is not removed. On querying status the
duplicate IO also appears closed but when a new fd is opened with
matching number the duplicate IO now points to this new, unrelated fd
that reused the number of the old closed fd. Now closing the stale
duplicte IO closes some unrelated fd causing failures in the
application.

To fix this instead of passing on fd numbers that are used with for_fd
pass a random cookie that is associated with the server side service and
look up the original IO from the service. Services are alrady tracked
anyway so just add the cookie and additional lookup by cookie.